pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "herald"
version = "0.1.0"
description = "An agentic email system that sends personalized stakeholder briefings"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "claude-agent-sdk>=0.1.48",
    "click>=8.0",
    "markdown>=3.5",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
databricks = ["databricks-sql-connector>=3.0"]
gmail = ["google-auth>=2.0", "google-auth-oauthlib>=1.0", "google-api-python-client>=2.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.3"]

[project.scripts]
herald = "herald.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/herald"]

[tool.ruff]
line-length = 100
target-version = "py310"
